# perf: Remove `Available()` call inside of filterInstanceTypesByRequirements