<!-- Please follow the guidelines at https://www.conventionalcommits.org/en/v1.0.0/ and use one of the following in your title:
feat:            <-- New features that require a MINOR version update
fix:             <-- Bug fixes that require at PATCH version update
chore:           <-- Smaller changes that impact behavior but aren't large enough to be features
perf:            <-- Code changes that improve performance but do not impact behavior
docs:            <-- Documentation changes that do not impact code
test:            <-- Test changes that do not impact behavior
ci:              <-- Changes that affect test or rollout automation
!${type}:        <-- Include ! if your change includes a backwards incompatible change.
-->

Fixes #N/A <!-- issue number -->

**Description**

This change ensures that we don't create a new slice every time we are evaluating the offerings for an instance type 

### Before PR

```
=== RUN   TestSchedulingProfile
scheduled 40151 against 8031 nodes in total in 1m0.598067383s 662.578886323755 pods/sec
400 instances 1 pods      1 nodes     181.661µs per scheduling      181.661µs per pod
400 instances 50 pods     10 nodes    43.569498ms per scheduling    871.389µs per pod
400 instances 100 pods    20 nodes    85.327955ms per scheduling    853.279µs per pod
400 instances 500 pods    100 nodes   420.066861ms per scheduling   840.133µs per pod
400 instances 1000 pods   200 nodes   801.483167ms per scheduling   801.483µs per pod
400 instances 1500 pods   300 nodes   1.213672s per scheduling      809.114µs per pod
400 instances 2000 pods   400 nodes   1.709432125s per scheduling   854.716µs per pod
400 instances 5000 pods   1000 nodes  5.00071875s per scheduling    1.000143ms per pod
400 instances 10000 pods  2000 nodes  12.3943185s per scheduling    1.239431ms per pod
400 instances 20000 pods  4000 nodes  37.195281583s per scheduling  1.859764ms per pod
--- PASS: TestSchedulingProfile (68.10s)
PASS
```

![heap](https://github.com/user-attachments/assets/799d5a77-a022-4ef5-82eb-f5c72064e89f)

### After PR

```
=== RUN   TestSchedulingProfile
scheduled 40151 against 8031 nodes in total in 56.201172872s 714.4156954063078 pods/sec
400 instances 1 pods      1 nodes     170.029µs per scheduling      170.029µs per pod
400 instances 50 pods     10 nodes    47.149229ms per scheduling    942.984µs per pod
400 instances 100 pods    20 nodes    79.660355ms per scheduling    796.603µs per pod
400 instances 500 pods    100 nodes   419.326583ms per scheduling   838.653µs per pod
400 instances 1000 pods   200 nodes   817.785854ms per scheduling   817.785µs per pod
400 instances 1500 pods   300 nodes   1.191372583s per scheduling   794.248µs per pod
400 instances 2000 pods   400 nodes   1.644508875s per scheduling   822.254µs per pod
400 instances 5000 pods   1000 nodes  4.748639875s per scheduling   949.727µs per pod
400 instances 10000 pods  2000 nodes  12.1352705s per scheduling    1.213527ms per pod
400 instances 20000 pods  4000 nodes  33.290514958s per scheduling  1.664525ms per pod
--- PASS: TestSchedulingProfile (63.97s)
PASS
```

![heap](https://github.com/user-attachments/assets/86707412-7774-42be-a574-86855da62068)

**How was this change tested?**

`make presubmit`

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.